% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delivery_list.R
\name{delivery_list}
\alias{delivery_list}
\title{delivery list}
\usage{
delivery_list(yelp_key, location, order)
}
\arguments{
\item{yelp_key}{a string representing the Yelp API key}

\item{location}{a string representing the delivery location, prefer zipcode}

\item{order}{a string representing the outcome order method}
}
\value{
a dataframe of max 20 restaurant that can deliver to a selected address based on selected order
}
\description{
Produces a dataframe of max 20 restaurant that can deliver to a select address ordered by review count/price/rating
}
\examples{
delivery_list(Sys.getenv("yelp_key"), "98104", "Rating")
}
\author{
Ruoqi xu, Apr 10
}
