% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_map.R
\name{location_map}
\alias{location_map}
\title{location map}
\usage{
location_map(yelp_key, category, city)
}
\arguments{
\item{yelp_key}{a string representing the Yelp API key}

\item{category}{a string representing the category of the business}

\item{city}{a string representing a city name}
}
\value{
a google map visualization of the top 5 business based on review counts
}
\description{
Produces a google map visualization of top 5 businesses (ordered by review count)
}
\examples{
location_map(Sys.getenv("yelp_key"), "cafe", "Vancouver")
}
\author{
Linsey Yao, Apr 10
}
