% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_match.R
\name{name_match}
\alias{name_match}
\title{name_match}
\usage{
name_match(key, name, city, state, country)
}
\arguments{
\item{key}{a string representing the Yelp API key}

\item{name}{a string representing the name of the business}

\item{city}{a string representing the city name}

\item{state}{a string representing the state code}

\item{country}{a string representing the country code}
}
\value{
a dataframe representing name, phone, location and postal code of a business that is the best match
}
\description{
Returns name, phone, location and postal code of a business that is the best match based on the name provided.
}
\examples{
name_match(Sys.getenv("yelp_key"), "Starbucks", "Burnaby", "BC", "CA")
}
\author{
Ying Dong, Apr 10
}
