% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cat.R
\name{search_cat}
\alias{search_cat}
\title{search_cat}
\usage{
search_cat(yelp_key, category, city)
}
\arguments{
\item{yelp_key}{a string representing the Yelp API key}

\item{category}{a string representing the category of the business}

\item{city}{a string representing a city name}
}
\value{
a dataframe containing business name, rating and review count
}
\description{
Produces a table containing names, ratings and review counts of 50 businesses on Yelp
}
\examples{
search_cat(Sys.getenv("yelp_key"), "cafe", "Vancouver")
}
