% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpitout-package.R
\docType{package}
\name{yelpitout}
\alias{yelpitout}
\alias{yelpitout-package}
\title{yelpitout: Search, delivery information and plot google map}
\description{
Search by category or name, delivery information and plot google map of different businesses on Yelp
}
